Clean up test utils
