+++ tests/utils/index.ts
@@ -1,0 +1,5 @@
+export * from "./bigint";
+export * from "./idl";
+export * from "./instructions";
+export * from "./misc";
+export * from "./spl";
